+++ martypy/Marty.py
@@ -15,7 +15,8 @@ my_marty.dance()
 The tags :one: and :two: indicate when the method is available for Marty V1 :one: and Marty V2 :two:
 '''
 import time
-from typing import Callable, Dict, List, Optional, Union
+from typing import Callable, Dict, List, Optional, Union, Tuple
+from enum import Enum
 from .ClientGeneric import ClientGeneric
 from .ClientMV2 import ClientMV2
 from .ClientMV1 import ClientMV1
@@ -82,6 +83,12 @@ class Marty(object):
     ADD_ON_TYPE_NAMES = [
         "IRFoot"
     ]
+
+    class Disco(Enum):
+        ARMS = {"00000088"}
+        FEET = {"00000087"}
+        EYES = {"00000089"}
+        ALL = {"00000087", "00000088", "00000089"}
 
     def __init__(self,
                 method: str,
@@ -155,7 +162,7 @@ class Marty(object):
         # Get Marty details
         self.client.start()
 
-    def dance(self, side: str = 'right', move_time: int = 4500, blocking: Optional[bool] = None) -> bool:
+    def dance(self, side: str = 'right', move_time: int = 3000, blocking: Optional[bool] = None) -> bool:
         '''
         Boogie, Marty! :one: :two:
         Args:
@@ -331,8 +338,9 @@ class Marty(object):
             amount: How much to lean. The defaults and the exact meaning is
                 different between Marty V1 and V2:
                 - :one: If not specified or `None`, `amount` defaults to `50` (no
-                        specific unit).
+                        specific unit). Limit is -60/60 in either direction.
                 - :two: If not specified or `None`, `amount` defaults to `29` degrees.
+                        Limit for foward and back is 45 and limit for left and right is 60.
             move_time: How long this movement should last, in milliseconds.
             blocking: Blocking mode override; whether to wait for physical movement to
                 finish before returning. Defaults to the value returned by `self.is_blocking()`.
@@ -371,8 +379,8 @@ class Marty(object):
         Play a named sound (Marty V2 :two:) or make a tone (Marty V1 :one:)
         Args:
             name_or_freq_start: name of the sound, e.g. 'excited' or 'no_way' :two:
-            name_or_freq_start: starting frequency, Hz :one:
-            freq_end:  ending frequency, Hz :one:
+            name_or_freq_start: starting frequency, min 20, max 20000, Hz :one:
+            freq_end:  ending frequency, min 20, max 20000, Hz :one:
             duration: milliseconds, maximum 5000 :one:
         Returns:
             True if Marty accepted the request
@@ -583,11 +591,15 @@ class Marty(object):
             jointIDNo = self.JOINT_IDS.get(joint_name_or_num, 0)
         return self.client.get_joint_status(jointIDNo)
 
-    def get_distance_sensor(self) -> float:
+    def get_distance_sensor(self) -> Union[int, float]:
         '''
         Get the latest value from the distance sensor :one: :two:
         Returns:
-            The distance sensor reading (will return 0 if no distance sensor is found)
+            The distance sensor reading. The meaning of the returned value is different 
+            between Marty V1 and V2:
+                - :one: Returns a raw distance sensor reading as a `float`.
+                - :two: Returns the distance in millimeters as `int`.
+            Both will return 0 if no distance sensor is found.
         '''
         return self.client.get_distance_sensor()
 
@@ -661,14 +673,19 @@ class Marty(object):
             none
         Returns:
             Dictionary containing:
-              "battRemainCapacityPercent" remaining battery capacity in percent
-              "battTempDegC" - battery temperature in degrees C
-              "battRemainCapacityMAH" - remaining battery capacity in milli-Amp-Hours
-              "battFullCapacityMAH" - capacity of the battery when full in milli-Amp-Hours
-              "battCurrentMA" - current the battery is supplying (or being charged with) milli-Amps
-              "power5VOnTimeSecs" - number of seconds the power to joints and add-ons has been on
-              "powerUSBIsConnected" - True if USB is connected
-              "power5VIsOn" - True if power to the joints and add-ons is turned on
+                "battRemainCapacityPercent": remaining battery capacity in percent
+                "battTempDegC": battery temperature in degrees C
+                "battRemainCapacityMAH": remaining battery capacity in milli-Amp-Hours
+                "battFullCapacityMAH": capacity of the battery when full in milli-Amp-Hours
+                "battCurrentMA": current the battery is supplying (or being charged with) milli-Amps
+                "power5VOnTimeSecs": number of seconds the power to joints and add-ons has been on
+                "powerUSBIsConnected": True if USB is connected
+                "power5VIsOn": True if power to the joints and add-ons is turned on
+
+                 Other values for internal use
+
+            **Note:** Some keys may not be included if Marty reports that the
+                      corresponding information is not available.
         '''
         return self.client.get_power_status()
     
@@ -703,21 +720,21 @@ class Marty(object):
         '''
         return self.client.get_add_on_status(add_on_name_or_id)
 
-    def add_on_query(self, addOnName: str, dataToWrite: bytes, numBytesToRead: int) -> Dict:
+    def add_on_query(self, add_on_name: str, data_to_write: bytes, num_bytes_to_read: int) -> Dict:
         '''
-        Write and read an addOn directly (raw-mode)
+        Write and read an add-on directly (raw-mode) :two:
         Args:
-            addOnName, name of the addOn (see get_add_ons_status() at the top level or response to
-                addon/list REST API command)
-            dataToWrite, can be zero length if nothing is to be written, the first byte will generally
-                be the register or opcode of the addOn
-            numBytesToRead: number of bytes to read from the device - can be zero
+            add_on_name: name of the add-on (see get_add_ons_status() at the top level or response to
+                `addon/list` REST API command)
+            data_to_write: can be zero length if nothing is to be written, the first byte will generally
+                be the register or opcode of the add-on
+            num_bytes_to_read: number of bytes to read from the device - can be zero
         Returns:
             Dict with keys including:
                 "rslt" - the result which will be "ok" if the query succeeded
-                "dataRead" - the data read from the addOn
+                "dataRead" - the data read from the add-on
         '''
-        return self.client.add_on_query(addOnName, dataToWrite, numBytesToRead)
+        return self.client.add_on_query(add_on_name, data_to_write, num_bytes_to_read)
 
     def get_system_info(self) -> Dict:
         '''
@@ -839,6 +856,51 @@ class Marty(object):
         '''
         return self.client.is_conn_ready()
         
+    def disco_off(self, add_on: Union[Disco, str] = Disco.ALL) -> bool:
+        '''
+        Turn disco add on LEDs off :two:
+        Args:
+            add_on: add on name of which the function applies to
+        Returns:
+            True if Marty accepted the request
+        '''
+        if type(add_on) is str:
+            return self.client.disco_off(add_on)
+        else:
+            return self.client.disco_group_operation(self.client.disco_off, add_on.value, {})
+
+    def disco_pattern(self, pattern: int, add_on: Union[Disco, str] = Disco.ALL) -> bool:
+        '''
+        Turn on a pattern of lights on the disco LED add on :two:
+        Args:
+            pattern: 1 or 2, pattern of lights that user wants to use
+            add_on: add on name of which the function applies to
+        Returns:
+            True if Marty accepted the request
+        '''
+        if type(add_on) is str:
+            return self.client.disco_pattern(pattern, add_on)
+        else:
+            return self.client.disco_group_operation(self.client.disco_pattern, add_on.value, {'pattern':pattern})
+
+    def disco_color(self, color: Union[str, Tuple[int, int, int]] = 'white', 
+                    add_on: Union[Disco, str] = Disco.ALL, 
+                    region: Union[int, str] = 'all') -> bool:
+        '''
+        Turn on disco add on LED lights to a specific color :two:
+        Args:
+            color: color to switch the LEDs to; takes in a hex code, RGB tuple of integers between 0-255, 
+                   or one of the built in colors: white, red, blue, yellow, green, teal, pink, purple, orange
+            add_on: add on name of which the function applies to
+            region: 0,1,2; region on the add on
+        Returns:
+            True if Marty accepted the request
+        '''
+        if type(add_on) is str:
+            return self.client.disco_color(color, add_on, region)
+        else:
+            return self.client.disco_group_operation(self.client.disco_color, add_on.value, {'color':color, 'region':region}) 
+
     ''' 
     ============================================================
     The following commands are for Marty V1 Only