+++ CHANGES.md
@@ -1,3 +1,10 @@
+Version 3.3.0 :
+- New functions for controlling the disco add-on lights
+- New distance sensor function for Marty V2 that returns the distance in mm
+
+Version 3.2.1 :
+- Default `movetime` parameter for the `dance()` has been changed to 3000ms instead of 4500ms
+
 Version 3.2.0 :
 - New `Marty.add_on_query` method provides low-level access to Marty's addons
 - New `Marty.is_conn_ready` method for checking if Marty is connected